+++ web/models/story.ex
@@ -10,7 +10,7 @@ defmodule News.Story do
     field :commands, :map, default: %{}
     field :attrs, :map, default: %{}
     field :meta, {:array, :string}, default: []
-    field :score, :integer
+    field :score, :float
     field :ip, Ecto.INET
     belongs_to :user, News.User
     has_many :taggings, News.Tagging
